src/django/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.0,<24.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0
